requirements.txt
openai
huggingface-hub
transformers
torch
python-dotenv
gradio
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
uvicorn
